samples/speech-to-speech/backend/requirements.txt
aws-cdk-lib==2.194.0
aws_cdk.aws_lambda_python_alpha==2.194.0a0
aws-cdk.lambda-layer-kubectl-v30==2.0.1
constructs>=10.0.0,<11.0.0
cdk_nag==2.35.81
